+++ test/test_lat_lng.dart
@@ -1,0 +1,6 @@
+class TestLatLng {
+  final double latitude;
+  final double longitude;
+
+  const TestLatLng(this.latitude, this.longitude);
+}